Creating a Dashboard componenet
