+++ client/src/components/LoginPage.js
@@ -1,21 +1,21 @@
-import React from 'react';
-import { connect } from 'react-redux';
-import { startLogin } from '../actions/auth.js';
+// import React from 'react';
+// import { connect } from 'react-redux';
+// import { startLogin } from '../actions/auth.js';
 
 
-export const LoginPage = ({ startLogin }) => (
-    <div className="box-layout">
-        <div className="box-layout__box">
-            <h1 className="box-layout__title">Boilerplate</h1>
-            <p>Tagline for app.</p>
-            <button className="button" onClick={startLogin}>Login with Google</button>
-        </div>
-    </div>
-);
+// export const LoginPage = ({ startLogin }) => (
+//     <div className="box-layout">
+//         <div className="box-layout__box">
+//             <h1 className="box-layout__title">Boilerplate</h1>
+//             <p>Tagline for app.</p>
+//             <button className="button" onClick={startLogin}>Login with Google</button>
+//         </div>
+//     </div>
+// );
 
-const mapDispatchToProps = (dispatch) =>({
-    startLogin: () => dispatch(startLogin())
-});
+// const mapDispatchToProps = (dispatch) =>({
+//     startLogin: () => dispatch(startLogin())
+// });
 
-export default connect(undefined, mapDispatchToProps)(LoginPage);
+// export default connect(undefined, mapDispatchToProps)(LoginPage);
