+++ client/src/routers/AppRouter.js
@@ -2,7 +2,9 @@ import React from "react";
 import ReactDOM from 'react-dom';
 import { Router, Route, Switch, Link, NavLink } from 'react-router-dom';
 import createHistory from 'history/createBrowserHistory';
+//This component is on the block to be deleted
 import LoginPage from '../components/LoginPage';
+//This component is on the block to be deleted
 import DashboardPage from '../components/DashboardPage';
 import NotFoundPage from '../components/NotFoundPage';
 import PrivateRoute from './PrivateRoute';
@@ -19,6 +21,7 @@ import NewWell from "../components/NewWell";
 import NewWellFull from "../components/NewWellFull";
 import csvParse from "../components/UploadCSV";
 import Graph3D from "../components/Graph3D";
+import Dashboard from "../components/Dashboard";
 
 export const history = createHistory();
 
@@ -29,7 +32,7 @@ const AppRouter = () => (
                 <PublicRoute path="/" component={Main} exact={true}/>
                 <PublicRoute path="/login" component={Login} exact={true} />
                 <PublicRoute path="/sign-up" component={Sign} exact={true} />
-                <PrivateRoute path="/dashboard" component={DashboardPage}/>
+                <PrivateRoute path="/dashboard" component={Dashboard}/>
                 <PrivateRoute path="/new-well-simplified" component={NewWell}/>
                 <PrivateRoute path="/my-wells" component={MyWells}/>
                 <PrivateRoute path="/account" component={Account}/>
